Refactored the code creating item classes.
Especially this change gets rid of the weird “queue” class level
attribute, which really is not needed. 

This also splits up and moves the Policy module below the Item class,
and makes sure these methods are already extended into Item.
