+++ lib/postqueue/item.rb
@@ -1,4 +1,7 @@
 require "active_record"
+require_relative "item/inserter"
+require_relative "item/policy"
+require_relative "item/migrations"
 
 module Postqueue
   #
@@ -7,23 +10,14 @@ module Postqueue
     self.table_name = nil
     self.abstract_class = true
 
-    def self.queue=(queue)
-      @queue = queue
+    def self.queue_support?
+      @queue_support = column_names.include?("queue") if @queue_support.nil?
+      @queue_support
     end
 
-    def self.queue
-      @queue
-    end
-
-    def self.create_item_class(queue:, table_name:)
+    def self.create_item_class(table_name:)
       klass = Class.new(self)
       klass.table_name = table_name
-      klass.queue = queue
-      klass.extend Postqueue::Policy
-
-      # We need to give this class a name, otherwise a number of AR operations
-      # are really really slow.
-      Postqueue::Item.const_set(dynamic_item_class_name, klass)
       klass
     end
 
@@ -31,7 +25,8 @@ module Postqueue
       @dynamic_item_class_count ||= 0
       "Dynamic#{@dynamic_item_class_count += 1}"
     end
+
+    extend Inserter
+    extend Policy
   end
 end
-
-require_relative "item/inserter"
