+++ lib/postqueue/item/inserter.rb
@@ -1,18 +1,11 @@
 require "active_record"
 
 module Postqueue
-  #
-  # Postqueue::Item inserter modules.
-  #
-  # This source file provides multiple implementations to insert Postqueue::Items.
-  # Which one will be used depends on the "extend XXXInserter" line below.
   class Item < ActiveRecord::Base
-    module RawInserter
-      def queue_support?
-        @queue_support = column_names.include?("queue") if @queue_support.nil?
-        @queue_support
-      end
-
+    module Inserter
+      # Inserter modules.
+      #
+      # This module provides a fast method to insert items
       def queue_attribute_names
         @queue_attribute_names ||= queue_support? ? [ :op, :entity_id, :queue ] : [ :op, :entity_id ]
       end
@@ -28,11 +21,10 @@ module Postqueue
         @insert_sql ||= begin
           columns = queue_attribute_names
           placeholders = 1.upto(columns.count).map { |i| "$#{i}" }
-          "INSERT INTO #{table_name}(#{columns.join(", ")}) VALUES(#{placeholders.join(", ")})"
+          quoted_table_name = connection.quote_table_name table_name
+          "INSERT INTO #{quoted_table_name}(#{columns.join(", ")}) VALUES(#{placeholders.join(", ")})"
         end
       end
     end
-
-    extend RawInserter # 200µs per item
   end
 end
